Added page refreshing for mobile and desktop platforms
When clicking show more or a post category, if the next page is liked
but the custom back button is not clicked, which calls javascript to
reload page, the new JS will do it from the native browser button. One
JS code block is for desktop, the other in mobile.